examples/terraform/aws/lambda-api-container/requirements.txt
boto3
psycopg2-binary
flask